requirements.txt
google-adk>=0.1.0
mcp>=0.1.0
litellm>=1.41.27